Rename oidc-broker and project content
To align all namings, we rename:
- the project itself, contaings code for multiple authd brokers to
  authd-oidc-brokers.
- oidc-broker implementation to authd-oidc.
This way, all projects relates to authd. We rename the snap names, URLs,
imports, configuration files to prevent stuttering.
