+++ internal/providers/withmsentraid.go
@@ -3,7 +3,7 @@
 package providers
 
 import (
-	"github.com/ubuntu/oidc-broker/internal/providers/microsoft_entra_id"
+	"github.com/ubuntu/authd-oidc-brokers/internal/providers/microsoft_entra_id"
 )
 
 // CurrentProviderInfo returns a Microsoft Entra ID provider implementation.
